requirements.txt
astroid==2.2.5
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
Django==2.1.7
django-crispy-forms==1.7.2
django-inspect==0.4.1
djangorestframework==3.9.2
gunicorn==19.9.0
idna==2.8
isort==4.3.16
lazy-object-proxy==1.3.1
mccabe==0.6.1
mysqlclient==1.4.2.post1
psycopg2-binary==2.8.2
PyJWT==1.7.1
pylint==2.3.1
pytz==2018.9
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.1
urllib3==1.24.2
virtualenv==16.4.3
wrapt==1.11.1
